requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
torch>=1.9.0
torch-geometric>=2.0.0
openpyxl>=3.0.0
scipy>=1.7.0
statsmodels>=0.12.0
xgboost>=1.4.0
pyyaml>=5.4.0